refactor: Fix style

+++ tasks/data_structures/build_heap.cpp
@@ -83,12 +83,12 @@ int main(int argc, char** argv) {
     if (argc == 1) {
         std::ios_base::sync_with_stdio(false);
 
-        unsigned n;
+        size_t n;
         std::cin >> n;
 
         std::vector<unsigned> data;
         data.resize(n);
-        for(unsigned i = 0; i < n; ++i)
+        for (size_t i = 0; i < n; ++i)
             std::cin >> data[i];
 
         auto swaps = solution(data);
